changed da converter data from 16 to 32 bit improved volume calculation with 32 bit precision

+++ main/display.h
@@ -6,5 +6,5 @@ void display_volume(uint8_t vol);
 void display_state(char *state, uint8_t *remote_name, uint8_t offset);
 void display_sample_rate(int sample_rate);
 void display_packets(uint32_t packets);
-void update_vu_meter(uint64_t level[2]);
+void update_vu_meter(uint32_t level[2]);
 void display_reboot();
